# Improve PCRE2 match performance for JIT and interpreted