--- a/src/regex/lib_pcre2.cr
+++ b/src/regex/lib_pcre2.cr
@@ -172,10 +172,10 @@ lib LibPCRE2
   CONFIG_COMPILED_WIDTHS   = 14
   CONFIG_TABLES_LENGTH     = 15
 
-  type Code = Void*
-  type CompileContext = Void*
-  type MatchData = Void*
-  type GeneralContext = Void*
+  type Code = Void
+  type CompileContext = Void
+  type MatchData = Void
+  type GeneralContext = Void
 
   fun get_error_message = pcre2_get_error_message_8(errorcode : Int, buffer : UInt8*, bufflen : LibC::SizeT) : Int
 
@@ -183,7 +183,7 @@ lib LibPCRE2
   fun code_free = pcre2_code_free_8(code : Code*) : Void
 
   type MatchContext = Void*
-  fun match_context_create = pcre2_match_context_create_8(gcontext : Void*) : MatchContext
+  fun match_context_create = pcre2_match_context_create_8(gcontext : Void*) : MatchContext*
 
   JIT_COMPLETE     = 0x00000001_u32 # For full matching
   JIT_PARTIAL_SOFT = 0x00000002_u32
@@ -191,15 +191,15 @@ lib LibPCRE2
   JIT_INVALID_UTF  = 0x00000100_u32
   fun jit_compile = pcre2_jit_compile_8(code : Code*, options : UInt32) : Int
 
-  type JITStack = Void*
+  type JITStack = Void
 
-  fun jit_stack_create = pcre2_jit_stack_create_8(startsize : LibC::SizeT, maxsize : LibC::SizeT, gcontext : GeneralContext) : JITStack
-  fun jit_stack_assign = pcre2_jit_stack_assign_8(mcontext : MatchContext, callable_function : Void* -> JITStack, callable_data : Void*) : Void
+  fun jit_stack_create = pcre2_jit_stack_create_8(startsize : LibC::SizeT, maxsize : LibC::SizeT, gcontext : GeneralContext*) : JITStack*
+  fun jit_stack_assign = pcre2_jit_stack_assign_8(mcontext : MatchContext*, callable_function : Void* -> JITStack*, callable_data : Void*) : Void
 
   fun pattern_info = pcre2_pattern_info_8(code : Code*, what : UInt32, where : Void*) : Int
 
-  fun match = pcre2_match_8(code : Code*, subject : UInt8*, length : LibC::SizeT, startoffset : LibC::SizeT, options : UInt32, match_data : MatchData*, mcontext : MatchContext) : Int
-  fun match_data_create_from_pattern = pcre2_match_data_create_from_pattern_8(code : Code*, gcontext : GeneralContext) : MatchData*
+  fun match = pcre2_match_8(code : Code*, subject : UInt8*, length : LibC::SizeT, startoffset : LibC::SizeT, options : UInt32, match_data : MatchData*, mcontext : MatchContext*) : Int
+  fun match_data_create_from_pattern = pcre2_match_data_create_from_pattern_8(code : Code*, gcontext : GeneralContext*) : MatchData*
   fun match_data_free = pcre2_match_data_free_8(match_data : MatchData*) : Void
 
   fun substring_nametable_scan = pcre2_substring_nametable_scan_8(code : Code*, name : UInt8*, first : UInt8*, last : UInt8*) : Int