This patch contains a number of individual steps that improve performance of PCRE2 matching greatly.

* Global `MatchContext` that is allocated only once and shared between all match executions. JIT stack is assigned via a callback which returns the appropriate thread-local stack. The callback won't be called if JIT is not used. This improve JIT performance.
* Cache `MatchData` per instance and thread. This avoids re-allocating the backtracking stack which improve interpreted performance
* Now that the pointer to `MatchData` does not leak outside the local scope, there's no need to allocate it with the GC. `libpcre2` can manage its memory on its own and the `MatchData` instances are handled in the bindings. This reduces GC pressure.

Using the benchmark program from https://github.com/crystal-lang/crystal/issues/13144#issue-1607857944, I get these results:

```console
$ bin/crystal run .test/bm-pcre2.cr --release             # (master)
starts_with? 284.80k (  3.51µs) (±18.92%)  20.1kB/op   7.43× slower
    matches?   2.12M (472.51ns) (±114.77%)    128B/op        fastest
$ bin/crystal run .test/bm-pcre2.cr --release             # (performance/pcre2-match_context)
starts_with?  18.26M ( 54.77ns) (±18.56%)  0.0B/op   1.52× slower
    matches?  27.67M ( 36.14ns) (±16.79%)  0.0B/op        fastest
$ bin/crystal run .test/bm-pcre2.cr --release -Duse_pcre1 # (master)
starts_with?  19.73M ( 50.67ns) (±14.41%)  16.0B/op   2.22× slower
    matches?  43.73M ( 22.87ns) (±17.42%)   0.0B/op        fastest
```

This shows a great improvement in match performance. The PCRE1 implementation is still significantly more performant in JIT mode (`matches?`).
A factor for this could be that the PCRE1 bindings are not thread safe. I'll leave investigation into this as a follow-up and consider the main regression as resolved.

Resolves #13144